packages/cmm-docs-mcp/src/cmm_docs/requirements.txt
fastmcp>=0.1.0
pymupdf>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
mistralai>=1.0.0
python-dotenv>=1.0.0
